requirements.txt
Flask==2.3.2
jsonschema==4.17.3
pymongo==4.3.3
PyJWT==2.7.0
waitress==2.1.2

# requirements.txt

# This file lists the Python dependencies required for your application.

Flask==2.3.2
jsonschema==4.17.3
pymongo==4.3.3
PyJWT==2.7.0
Gunicorn==21.2.0

